Moving towards showing rule text in each view.
Added rule class & method to pull from DB. Also pulling rules as a tree
from the start, will be passing the “subRule” array to the child view
instead of hitting the DB again.

Tweaked the Section and Rule classes to init with a DB row instead of
creating params for every value. Mostly for clean code.